+++ example/moneyTransfer.js
@@ -1,21 +1,21 @@
 /**
  * 現金転送実行
- * @ignore
  */
+const mongoose = require('mongoose');
 
 const pecorino = require('../');
 
-pecorino.mongoose.connect(process.env.MONGOLAB_URI)
+mongoose.connect(process.env.MONGOLAB_URI)
     .then(async () => {
-        const taskRepo = new pecorino.repository.Task(pecorino.mongoose.connection);
+        const taskRepo = new pecorino.repository.Task(mongoose.connection);
 
         try {
             await pecorino.service.task.executeByName(
                 pecorino.factory.taskName.MoneyTransfer
-            )({ taskRepo: taskRepo, connection: pecorino.mongoose.connection });
+            )({ taskRepo: taskRepo, connection: mongoose.connection });
         } catch (error) {
             console.error(error.message);
         }
 
-        await pecorino.mongoose.disconnect();
+        await mongoose.disconnect();
     });