+++ src/repo/action.ts
@@ -29,14 +29,32 @@ export class MongoRepository {
         andConditions.push({
             $or: [
                 {
-                    'fromLocation.typeOf': factory.account.TypeOf.Account,
-                    'fromLocation.accountType': params.accountType,
-                    'fromLocation.accountNumber': params.accountNumber
+                    'fromLocation.typeOf': {
+                        $exists: true,
+                        $eq: factory.account.TypeOf.Account
+                    },
+                    'fromLocation.accountType': {
+                        $exists: true,
+                        $eq: params.accountType
+                    },
+                    'fromLocation.accountNumber': {
+                        $exists: true,
+                        $eq: params.accountNumber
+                    }
                 },
                 {
-                    'toLocation.typeOf': factory.account.TypeOf.Account,
-                    'toLocation.accountType': params.accountType,
-                    'toLocation.accountNumber': params.accountNumber
+                    'toLocation.typeOf': {
+                        $exists: true,
+                        $eq: factory.account.TypeOf.Account
+                    },
+                    'toLocation.accountType': {
+                        $exists: true,
+                        $eq: params.accountType
+                    },
+                    'toLocation.accountNumber': {
+                        $exists: true,
+                        $eq: params.accountNumber
+                    }
                 }
             ]
         });
@@ -51,9 +69,10 @@ export class MongoRepository {
             ...params,
             actionStatus: factory.actionStatusType.ActiveActionStatus,
             startDate: new Date()
-        }).then(
-            (doc) => doc.toObject()
-        );
+        })
+            .then(
+                (doc) => doc.toObject()
+            );
     }
     /**
      * アクション完了
@@ -74,13 +93,15 @@ export class MongoRepository {
                 endDate: new Date()
             },
             { new: true }
-        ).exec().then((doc) => {
-            if (doc === null) {
-                throw new factory.errors.NotFound('action');
-            }
+        )
+            .exec()
+            .then((doc) => {
+                if (doc === null) {
+                    throw new factory.errors.NotFound('action');
+                }
 
-            return doc.toObject();
-        });
+                return doc.toObject();
+            });
     }
     /**
      * アクション中止
@@ -96,7 +117,8 @@ export class MongoRepository {
             },
             { actionStatus: factory.actionStatusType.CanceledActionStatus },
             { new: true }
-        ).exec()
+        )
+            .exec()
             .then((doc) => {
                 if (doc === null) {
                     throw new factory.errors.NotFound('action');
@@ -125,16 +147,18 @@ export class MongoRepository {
                 endDate: new Date()
             },
             { new: true }
-        ).exec().then((doc) => {
-            if (doc === null) {
-                throw new factory.errors.NotFound('action');
-            }
+        )
+            .exec()
+            .then((doc) => {
+                if (doc === null) {
+                    throw new factory.errors.NotFound('action');
+                }
 
-            return doc.toObject();
-        });
+                return doc.toObject();
+            });
     }
     /**
-     * IDで取得する
+     * アクション検索
      */
     public async findById<T extends factory.actionType>(
         typeOf: T,
@@ -145,7 +169,8 @@ export class MongoRepository {
                 typeOf: typeOf,
                 _id: actionId
             }
-        ).exec()
+        )
+            .exec()
             .then((doc) => {
                 if (doc === null) {
                     throw new factory.errors.NotFound('action');
@@ -159,7 +184,9 @@ export class MongoRepository {
     ): Promise<number> {
         const conditions = MongoRepository.CREATE_MONEY_TRANSFER_ACTIONS_MONGO_CONDITIONS(params);
 
-        return this.actionModel.countDocuments({ $and: conditions }).setOptions({ maxTimeMS: 10000 }).exec();
+        return this.actionModel.countDocuments({ $and: conditions })
+            .setOptions({ maxTimeMS: 10000 })
+            .exec();
     }
     /**
      * 転送アクションを検索する
@@ -179,7 +206,8 @@ export class MongoRepository {
         // tslint:disable-next-line:no-single-line-block-comment
         /* istanbul ignore else */
         if (params.limit !== undefined && params.page !== undefined) {
-            query.limit(params.limit).skip(params.limit * (params.page - 1));
+            query.limit(params.limit)
+                .skip(params.limit * (params.page - 1));
         }
         // tslint:disable-next-line:no-single-line-block-comment
         /* istanbul ignore else */
@@ -187,7 +215,9 @@ export class MongoRepository {
             query.sort(params.sort);
         }
 
-        return query.setOptions({ maxTimeMS: 10000 }).exec().then((docs) => docs.map((doc) => doc.toObject()));
+        return query.setOptions({ maxTimeMS: 10000 })
+            .exec()
+            .then((docs) => docs.map((doc) => doc.toObject()));
     }
     /**
      * アクションを検索する
@@ -265,7 +295,6 @@ export class MongoRepository {
                 updatedAt: 0
             }
         )
-            .sort({ _id: 1 })
             .limit(searchConditions.limit)
             .exec()
             .then((docs) => docs.map((doc) => doc.toObject()));