+++ src/service/transaction/deposit.ts
@@ -36,8 +36,8 @@ export function start<T extends factory.account.AccountType>(
 
         // 口座存在確認
         const account = await repos.account.findByAccountNumber<T>({
-            accountType: params.object.accountType,
-            accountNumber: params.object.toAccountNumber
+            accountType: params.object.toLocation.accountType,
+            accountNumber: params.object.toLocation.accountNumber
         });
 
         // 取引ファクトリーで新しい進行中取引オブジェクトを作成
@@ -48,9 +48,13 @@ export function start<T extends factory.account.AccountType>(
             object: {
                 clientUser: params.object.clientUser,
                 amount: params.object.amount,
-                accountType: params.object.accountType,
-                toAccountNumber: account.accountNumber,
-                notes: params.object.notes
+                toLocation: {
+                    typeOf: factory.account.TypeOf.Account,
+                    accountType: account.accountType,
+                    accountNumber: account.accountNumber,
+                    name: account.name
+                },
+                description: params.object.description
             },
             expires: params.expires
         };
@@ -77,8 +81,8 @@ export function start<T extends factory.account.AccountType>(
 
         // 入金先口座に進行中取引を追加
         await repos.account.startTransaction<T>({
-            accountType: params.object.accountType,
-            accountNumber: params.object.toAccountNumber,
+            accountType: params.object.toLocation.accountType,
+            accountNumber: params.object.toLocation.accountNumber,
             transaction: pendingTransaction
         });
 
@@ -104,7 +108,7 @@ export function confirm<T extends factory.account.AccountType>(params: {
         // 現金転送アクション属性作成
         const moneyTransferActionAttributes: factory.action.transfer.moneyTransfer.IAttributes<T> = {
             typeOf: factory.actionType.MoneyTransfer,
-            description: transaction.object.notes,
+            description: transaction.object.description,
             result: {
                 amount: transaction.object.amount
             },
@@ -113,14 +117,14 @@ export function confirm<T extends factory.account.AccountType>(params: {
             agent: transaction.agent,
             recipient: transaction.recipient,
             amount: transaction.object.amount,
-            fromLocation: {
-                typeOf: transaction.agent.typeOf,
-                name: transaction.agent.name
-            },
+            fromLocation: transaction.object.fromLocation !== undefined
+                ? transaction.object.fromLocation
+                : /* istanbul ignore next */ {
+                    typeOf: transaction.agent.typeOf,
+                    name: transaction.agent.name
+                },
             toLocation: {
-                typeOf: factory.account.TypeOf.Account,
-                accountType: transaction.object.accountType,
-                accountNumber: transaction.object.toAccountNumber,
+                ...transaction.object.toLocation,
                 name: transaction.recipient.name
             },
             purpose: {
@@ -158,7 +162,7 @@ export function exportTasks(status: factory.transactionStatusType) {
 }
 
 /**
- * ID指定で取引のタスク出力
+ * 取引のタスク出力
  */
 export function exportTasksById<T extends factory.account.AccountType>(
     transactionId: string
@@ -186,7 +190,6 @@ export function exportTasksById<T extends factory.account.AccountType>(
                             status: factory.taskStatus.Ready,
                             runsAt: new Date(), // なるはやで実行
                             remainingNumberOfTries: 10,
-                            lastTriedAt: null,
                             numberOfTried: 0,
                             executionResults: [],
                             data: {
@@ -205,7 +208,6 @@ export function exportTasksById<T extends factory.account.AccountType>(
                     status: factory.taskStatus.Ready,
                     runsAt: new Date(), // なるはやで実行
                     remainingNumberOfTries: 10,
-                    lastTriedAt: null,
                     numberOfTried: 0,
                     executionResults: [],
                     data: {